Refactor enums with explicit types; update project configs
Refactored enums across the codebase to use explicit underlying types (mainly uint8_t/int8_t) for improved type safety and clarity, updating all related usages accordingly. Changed some constants to constexpr and updated default enum initializations. Added <cstdint> includes where needed. Updated Mesh.cpp to use new enum scoping. Revised all .vcxproj files to use absolute vcpkg/Bullet include and lib paths, and fixed minor formatting. Added VSCode launch and settings JSON files for improved C/C++ development and debugging support. These changes modernize enum usage and enhance build configuration robustness.

+++ Tractor3Dlib/src/framework/Platform.cpp
@@ -114,7 +114,7 @@ void Platform::gamepadButtonPressedEventInternal(GamepadHandle handle, Gamepad::
     Gamepad* gamepad = Gamepad::getGamepad(handle);
     if (gamepad)
     {
-        unsigned int newButtons = gamepad->_buttons | (1 << mapping);
+        unsigned int newButtons = gamepad->_buttons | (1 << static_cast<uint8_t>(mapping));
         gamepad->setButtons(newButtons);
         Form::gamepadButtonEventInternal(gamepad);
     }
@@ -126,7 +126,7 @@ void Platform::gamepadButtonReleasedEventInternal(GamepadHandle handle, Gamepad:
     Gamepad* gamepad = Gamepad::getGamepad(handle);
     if (gamepad)
     {
-        unsigned int newButtons = gamepad->_buttons & ~(1 << mapping);
+        unsigned int newButtons = gamepad->_buttons & ~(1 << static_cast<uint8_t>(mapping));
         gamepad->setButtons(newButtons);
         Form::gamepadButtonEventInternal(gamepad);
     }
